refactoring track view models for further improvements
