--- app/models/tracks/raw_track_presenter.rb
@@ -1,9 +1,0 @@
-module Tracks
-  class RawTrackPresenter < TrackPresenter
-    protected
-
-    def preprocess_points(raw_points)
-      PointsPostprocessor::Default.new(raw_points).execute
-    end
-  end
-end
