+++ config/locales/models/tournament_matches/en.yml
@@ -1,0 +1,4 @@
+en:
+  activerecord:
+    models:
+      tournament_match: Bracket
