+++ config/locales/models/tournament_matches/fr.yml
@@ -1,0 +1,4 @@
+fr:
+  activerecord:
+    models:
+      tournament_match: Match
