+++ config/locales/models/tournament_matches/it.yml
@@ -1,0 +1,4 @@
+it:
+  activerecord:
+    models:
+      tournament_match: Staffa
