+++ config/locales/models/tournament_matches/ru.yml
@@ -1,0 +1,4 @@
+ru:
+  activerecord:
+    models:
+      tournament_match: Пара
